pyproject.toml
[project]
name = "gdpr-agent"
version = "0.1.0"
description = "GDPR Legal Analyst Agent — incremental GCP AI/ML learning project"
requires-python = ">=3.11"
dependencies = [
    # LangChain core + GCP integrations
    "langchain>=0.3",
    "langchain-text-splitters>=0.3",
    "langchain-google-vertexai>=2.0",
    "langchain-google-community[bigquery]>=2.0",
    "langchain-community>=0.3",
    # GCP SDKs
    "google-cloud-aiplatform>=1.60",
    "google-cloud-bigquery>=3.20",
    # Document loading
    "pypdf>=4.0",
    # Config & CLI
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "typer>=0.12",
]

[project.optional-dependencies]
dev = [
    "ipython",
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["phase1", "setup", "config.py"]
